3711: Added step create action

+++ app/models/project_step.rb
@@ -221,14 +221,16 @@ class ProjectStep < ActiveRecord::Base
   # todo: refactor up to translatable.rb
 
   def update_translations!(translation_params)
-    # deleting the translations that have been removed
-    translation_params[:deleted_locales].each do |l|
-      [:details, :summary].each do |attr|
-        delete_translation(attr, l)
+    if persisted?
+      # deleting the translations that have been removed
+      translation_params[:deleted_locales].each do |l|
+        [:details, :summary].each do |attr|
+          delete_translation(attr, l)
+        end
       end
-    end
 
-    reload
+      reload
+    end
 
     # updating/creating the translation that have been updated, added
     permitted_locales.each do |l|
